fix(storage): canonicalize ledger state, not only its views
Two producers declaring identical records in different orders built ledgers
that compared equal and serialized identically while their `repr`s differed,
so a log line or diagnostic depended on collection order.

This is the third time this branch has broken the same invariant, and the
invariant is written down in this package's own AGENTS.md from the first
time: "a claim about the stored state, not only about accessors: a canonical
view over non-canonical state leaves `__eq__` and `repr` exposed."
`OccurrenceSet` kept insertion order behind a sorted `__iter__`;
`StorageVersions` normalized only in `to_dict`; here `__eq__` agreed only
because dict equality ignores key order, so `repr` was the one view with
nothing to hide behind. For a value type: canonicalize the state.

Sorted at the assignment door that already copies these mappings, so there
is one rule in one place. `declare`/`override` now reassign rather than
writing in place — an in-place write appends at the end and never reaches
that door, which is how the first version of this fix silently did nothing.

Swept the package's other stateful containers rather than fixing what was
reported: `OccurrenceSet`, `IdentityConflict` and
`StorageVersions.section_schema_versions` each already produce one state per
permutation, so this was the last member, and a test says so. Coverage is a
permutation property rather than two hand-picked orders — both previous
instances were fixed for the one ordering that had been demonstrated.

Two size caps fell out, both split along the dependency direction already
used for `entity_ids.py`: `FactAvailability` moves to `fact_availability.py`
(the record knows nothing about the ledger indexing it) and the new tests to
`test_ledger_state.py`. A test that reached the status sets through
`availability`'s alias now imports them from `availability_status`, which
owns them — importing through a module that merely re-exports is how a test
ends up pinned to a forwarding detail instead of to the rule.
